Removed phase adjustments
It turns out that the phase of a triangle wave can't be created from cartesian
form, so the real/imaginary representation is not helpful. (Run triangle.m and
observe that the waveforms summed from two near-identical triangle waves are
not triangular but have truncated peaks)

Also modified Makefile so it's easier to create more testbenches.